Revert "Bug/#100 test fixes"
