+++ .husky/pre-commit
@@ -1,2 +1,4 @@
-#!/usr/bin/env sh 
- 
+#!/usr/bin/env sh
+. "$(dirname -- "$0")/_/husky.sh"
+
+npx lint-staged
