+++ src/components/__tests__/GroupBookingForm.test.tsx
@@ -380,10 +380,10 @@ describe("GroupBookingForm", () => {
     });
 
     // Submit the form
-    // const submitButton = screen.getByRole("button", { name: "Submit Booking" });
-    // await act(async () => {
-    //   fireEvent.click(submitButton);
-    // });
+    const submitButton = screen.getByRole("button", { name: "Submit Booking" });
+    await act(async () => {
+      fireEvent.click(submitButton);
+    });
 
     // Wait for all async operations to complete
     await waitFor(() => {
@@ -452,10 +452,10 @@ describe("GroupBookingForm", () => {
     });
 
     // Submit the form
-    // const submitButton = screen.getByRole("button", { name: "Submit Booking" });
-    // await act(async () => {
-    //   fireEvent.click(submitButton);
-    // });
+    const submitButton = screen.getByRole("button", { name: "Submit Booking" });
+    await act(async () => {
+      fireEvent.click(submitButton);
+    });
 
     // Check if error is handled
     await waitFor(() => {
